Added links to Casual and Casual Suite (#517)
Casual is a collection of Transient user interfaces for various built-in Emacs
modes. This commit adds multiple links from Casual to highlight its offering to
a particular section.

Support for third party (non built-in) Emacs modes is handled with a separate
repo/package. Examples of this include Casual Avy and Casual Symbol Overlay.

Casual Suite is an umbrella package that includes all Casual user interfaces,
both built-in and third party.